+++ i2c.c
@@ -16,9 +16,9 @@
  */
 /* ************************************************************************** */
 
-#include "common.h"
+#include "pic32mx.h"
 #include "gpio.h"
-#include "i2c.h"
+#include "mrubyc.h"
 
 /* ================================ C codes ================================ */
 #define I2CFREQ 100000	// 100kHz
