--- i2c.h
@@ -1,34 +1,0 @@
-/* ************************************************************************** */
-/** I2C
-
-  @Company
-    ShimaneJohoshoriCenter.inc
-
-  @File Name
-    i2c.h
-
-  @Summary
-    I2C processing
-
-  @Description
-    mruby/c function army
- */
-/* ************************************************************************** */
-
-#ifndef RBOARD_I2C_H
-#define RBOARD_I2C_H
-
-/* Provide C++ Compatibility */
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/* mruby/c codes */
-void mrbc_init_class_i2c(void);
-
-/* Provide C++ Compatibility */
-#ifdef __cplusplus
-}
-#endif
-
-#endif /* RBOARD_I2C_H */
