+++ nbproject/configurations.xml
@@ -4,13 +4,6 @@
     <logicalFolder name="HeaderFiles"
                    displayName="Header Files"
                    projectFiles="true">
-      <itemPath>adc.h</itemPath>
-      <itemPath>delay.h</itemPath>
-      <itemPath>digital.h</itemPath>
-      <itemPath>i2c.h</itemPath>
-      <itemPath>mrbc_firm.h</itemPath>
-      <itemPath>spi.h</itemPath>
-      <itemPath>uart.h</itemPath>
     </logicalFolder>
     <logicalFolder name="LinkerScript"
                    displayName="Linker Files"
@@ -40,14 +33,15 @@
         <itemPath>src/error.c</itemPath>
         <itemPath>src/c_object.c</itemPath>
       </logicalFolder>
-      <itemPath>delay.c</itemPath>
       <itemPath>main.c</itemPath>
       <itemPath>i2c.c</itemPath>
       <itemPath>uart.c</itemPath>
       <itemPath>mrbc_firm.c</itemPath>
       <itemPath>adc.c</itemPath>
-      <itemPath>digital.c</itemPath>
       <itemPath>spi.c</itemPath>
+      <itemPath>gpio.c</itemPath>
+      <itemPath>pic32mx.c</itemPath>
+      <itemPath>pwm.c</itemPath>
     </logicalFolder>
     <logicalFolder name="ExternalFiles"
                    displayName="Important Files"
@@ -109,7 +103,7 @@
         <property key="enable-symbols" value="true"/>
         <property key="enable-unroll-loops" value="false"/>
         <property key="exclude-floating-point" value="false"/>
-        <property key="extra-include-directories" value=".;src"/>
+        <property key="extra-include-directories" value=".;PIC32MX170F256B;src"/>
         <property key="generate-16-bit-code" value="false"/>
         <property key="generate-micro-compressed-code" value="false"/>
         <property key="isolate-each-function" value="false"/>
