+++ pic32mx.h
@@ -1,16 +1,22 @@
-#ifndef COMMON_H
-#define COMMON_H
+/*
+  PIC32MX Related functions.
+*/
 
+#ifndef PIC32MX_H
+#define PIC32MX_H
+
+#include <xc.h>
 #include <stdint.h>
-#include "pic32mx.h"
-#include "mrubyc.h"
-
+#include "model_dependent.h"
 
 /* Provide C++ Compatibility */
 #ifdef __cplusplus
 extern "C" {
 #endif
 
+/*! Pin handle struct.
+
+*/
 typedef struct PIN_HANDLE {
   uint8_t port;		// A=1,B=2,..,G=7
   uint8_t num;		// 0..15
@@ -22,7 +28,9 @@ void __delay_ms(uint32_t ms);
 void system_register_lock(void);
 void system_register_unlock(void);
 void system_reset(void);
-int set_pin_handle( PIN_HANDLE *pin_handle, const mrbc_value *val );
+
+struct RObject;
+int set_pin_handle( PIN_HANDLE *pin_handle, const struct RObject *val );
 
 
 /* Provide C++ Compatibility */
@@ -31,4 +39,4 @@ int set_pin_handle( PIN_HANDLE *pin_handle, const mrbc_value *val );
 #endif
 
 
-#endif /* COMMON_H */
+#endif /* PIC32MX_H */
