--- pwm.h
@@ -1,37 +1,0 @@
-/* ************************************************************************** */
-/** PWM
-
-  @Company
-    ShimaneJohoshoriCenter.inc
-
-  @File Name
-    pwm.h
-
-  @Summary
-    PWM processing
-
-  @Description
-    mruby/c function army
- */
-/* ************************************************************************** */
-
-#ifndef RBOARD_PWM_H
-#define RBOARD_PWM_H
-
-
-/* Provide C++ Compatibility */
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/* C codes */
-
-/* mruby/c codes */
-void mrbc_init_class_pwm(void);
-
-/* Provide C++ Compatibility */
-#ifdef __cplusplus
-}
-#endif
-
-#endif /* RBOARD_PWM_H */
