+++ spi.c
@@ -15,8 +15,9 @@
  */
 /* ************************************************************************** */
 
-#include "common.h"
+#include "pic32mx.h"
 #include "gpio.h"
+#include "mrubyc.h"
 
 /* ================================ C codes ================================ */
