--- spi.h
@@ -1,34 +1,0 @@
-/* ************************************************************************** */
-/** SPI
-
-  @Company
-    ShimaneJohoshoriCenter.inc
-
-  @File Name
-    spi.h
-
-  @Summary
-    SPI processing
-
-  @Description
-    mruby/c function army
- */
-/* ************************************************************************** */
-
-#ifndef RBOARD_SPI_H
-#define RBOARD_SPI_H
-
-/* Provide C++ Compatibility */
-#ifdef __cplusplus
-extern "C" {
-#endif
-
-/* mruby/c codes */
-void mrbc_init_class_spi(void);
-
-    /* Provide C++ Compatibility */
-#ifdef __cplusplus
-}
-#endif
-
-#endif /* RBOARD_SPI_H */
